+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.5.2 (unreleased)
 
+  - Added "IdentitiesOnly yes" to options when `vagrant ssh` is run to
+    avoid "Too Many Authentication Failures" error. [GH-131]
   - Fix regression with `package` not working. [GH-132]
   - Added ability to specify box url in `init`, which populates the
     Vagrantfile with the proper `config.vm.box_url`.